Move code out that is responsible for importing data into the datastore database in preparation for the samples.
Ensure query profiling output in the script matches what the sample tests expect.
