+++ samples/test/queryFilterOr.test.js
@@ -16,15 +16,12 @@
 
 'use strict';
 
-const path = require('path');
 const {assert} = require('chai');
-const {describe, it, after, before} = require('mocha');
+const {describe, it} = require('mocha');
 const sinon = require('sinon');
-const {Datastore, PropertyFilter, or} = require('@google-cloud/datastore');
-const datastore = new Datastore();
+// TODO: delete unused imports
 
 const cp = require('child_process');
-let taskKey1, taskKey2;
 
 const execSync = cmd => cp.execSync(cmd, {encoding: 'utf-8'});
 
@@ -42,43 +39,7 @@ describe('Creating a union query', () => {
   beforeEach(stubConsole);
   afterEach(restoreConsole);
 
-  before(async () => {
-    taskKey1 = datastore.key('Task');
-    const entity1 = {
-      key: taskKey1,
-      data: {
-        description: 'Buy milk',
-      },
-    };
-
-    taskKey2 = datastore.key('Task');
-    const entity2 = {
-      key: taskKey2,
-      data: {
-        description: 'Feed cats',
-      },
-    };
-
-    // Ensure the datastore database has no existing data for the task keys.
-    const query = datastore
-      .createQuery('Task')
-      .filter(
-        or([
-          new PropertyFilter('description', '=', 'Buy milk'),
-          new PropertyFilter('description', '=', 'Feed cats'),
-        ])
-      );
-    const [entities] = await datastore.runQuery(query);
-    await datastore.delete(entities.map(entity => entity[datastore.KEY]));
-
-    await datastore.upsert(entity1);
-    await datastore.upsert(entity2);
-  });
-
-  after(async () => {
-    await datastore.delete(taskKey1);
-    await datastore.delete(taskKey2);
-  });
+  require('./helpers/populate-data');
 
   it('should get a combination of items from the Datastore', async () => {
     assert.strictEqual(execSync(`node ./queryFilterOr.js`), 'Entity found: Buy milk\nEntity found: Feed cats\n');
